fix(mcast): first-compile fixes for the cloud-authored mesh stack
The mesh-voice branch was authored on a runner with no ATAK SDK, so
this is its first real compile + test run. Two classes of break:

- XvMapComponent passed MumbleTransport's Long speakerSession straight
  into MeshVoiceManager.onMumbleRxFrame(Int). Narrow at the call site;
  Int is the codebase-wide session type (presence registry, CoT
  publisher, roster) — the transport callback Long is the outlier.
- ChannelMulticastConfigTest / CommsPlanTest / CommsPlanCarrierTest ran
  as plain JUnit but exercise org.json, which the mockable android.jar
  stubs to return-default-values — JSONObject.quote silently emits
  nulls. Run them under Robolectric, same pattern as the existing
  XvSettings tests.

+++ app/src/main/java/com/atakmap/android/xv/plugin/XvMapComponent.kt
@@ -4357,7 +4357,7 @@ class XvMapComponent : AbstractMapComponent() {
                     // A true return (or no mesh manager) plays as before.
                     val play =
                         try {
-                            meshVoiceManager?.onMumbleRxFrame(slot, speakerSession, opus) ?: true
+                            meshVoiceManager?.onMumbleRxFrame(slot, speakerSession.toInt(), opus) ?: true
                         } catch (t: Throwable) {
                             Log.w(TAG, "mesh onMumbleRxFrame threw", t)
                             true
